# docs/swagger.yaml
definitions:
  controllers.DeleteReponse:
    properties:
      message:
        type: string
    type: object
  controllers.EmailResponse:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
  controllers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  controllers.LoginResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  controllers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  controllers.RegisterResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  controllers.UserReponse:
    properties:
      message:
        type: string
    type: object
  controllers.UserRequest:
    properties:
      email:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/DeleteUser/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user given their ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.DeleteReponse'
      security:
      - BearerAuth: []
      tags:
      - User
  /api/v1/GetCurrentUser:
    get:
      consumes:
      - application/json
      description: Get current user by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.EmailResponse'
      security:
      - BearerAuth: []
      tags:
      - User
  /api/v1/UpdateUser/{id}:
    put:
      consumes:
      - application/json
      description: Update a user's email
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UserReponse'
      security:
      - BearerAuth: []
      tags:
      - User
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login Request
        in: body
        name: request
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.LoginResponse'
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register Request
        in: body
        name: request
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.RegisterResponse'
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"